% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosechar.R
\name{cargar}
\alias{cargar}
\title{Cargar Objeto SeMiLLa}
\usage{
cargar(archivo)
}
\arguments{
\item{archivo}{Ruta al archivo \code{.rds}.}
}
\value{
El objeto guardado en el archivo (normalmente de clase
\code{semilla}), tal como lo devuelve \code{readRDS()}.
}
\description{
Carga un objeto SeMiLLa guardado previamente con \code{\link{guardar}}.
}
\examples{
f <- file.path(tempdir(), "semilla_demo")
ruta <- guardar(semilla_demo, f, verbose = FALSE)
x <- cargar(ruta)
class(x)
unlink(ruta)

}
