% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hacer_definida_positiva}
\alias{.hacer_definida_positiva}
\title{Hacer una matriz definida positiva}
\usage{
.hacer_definida_positiva(mat, tol = 1e-06)
}
\description{
Hacer una matriz definida positiva
}
\keyword{internal}
