% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensamblar.R
\name{.img_dims_inch}
\alias{.img_dims_inch}
\title{Calcula el ancho/alto en pulgadas conservando la relacion de aspecto.}
\usage{
.img_dims_inch(ruta, ancho_obj = 4.5, alto_max = 6)
}
\description{
Calcula el ancho/alto en pulgadas conservando la relacion de aspecto.
}
\keyword{internal}
