% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.transformar_similitud_para_cfa}
\alias{.transformar_similitud_para_cfa}
\title{Transformar matriz de similitud semantica para CFA}
\usage{
.transformar_similitud_para_cfa(
  similitud,
  metodo = "semantico",
  verbose = TRUE
)
}
\description{
Aplica transformaciones para que la matriz de similitud semantica
se comporte mas como una matriz de correlacion empirica.
}
\keyword{internal}
