% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asignacion.R
\name{print.semilla_asignacion}
\alias{print.semilla_asignacion}
\title{Imprimir la auditoria de asignacion}
\usage{
\method{print}{semilla_asignacion}(x, ...)
}
\arguments{
\item{x}{Objeto \code{semilla_asignacion}.}

\item{...}{Sin uso.}
}
\value{
El objeto, invisible.
}
\description{
Imprimir la auditoria de asignacion
}
