% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_semantico_archivo.R
\name{print.semilla_efa}
\alias{print.semilla_efa}
\title{Imprimir EFA (ARCHIVADO)}
\usage{
\method{print}{semilla_efa}(x, ...)
}
\arguments{
\item{x}{Objeto semilla_efa}

\item{...}{Argumentos adicionales}
}
\description{
Imprimir EFA (ARCHIVADO)
}
