% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancia.R
\name{print.semilla_redundancia}
\alias{print.semilla_redundancia}
\title{Imprimir auditoria de redundancia}
\usage{
\method{print}{semilla_redundancia}(x, ...)
}
\arguments{
\item{x}{Objeto \code{semilla_redundancia}.}

\item{...}{Ignorado.}
}
\value{
El objeto \code{x} de forma invisible.
}
\description{
Imprimir auditoria de redundancia
}
