% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analizar_tolerancia.R
\name{print.semilla_tolerancia}
\alias{print.semilla_tolerancia}
\title{Imprime el resumen de un analisis de tolerancia}
\usage{
\method{print}{semilla_tolerancia}(x, ...)
}
\arguments{
\item{x}{Objeto \code{semilla_tolerancia} de \code{\link{analizar_tolerancia}}.}

\item{...}{Ignorado.}
}
\value{
Devuelve \code{x} de forma invisible.
}
\description{
Imprime el resumen de un analisis de tolerancia
}
