% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosechar.R
\name{cargar}
\alias{cargar}
\title{Cargar Objeto SeMiLLa}
\usage{
cargar(archivo)
}
\arguments{
\item{archivo}{Ruta al archivo .rds}
}
\value{
Objeto semilla
}
\description{
Carga un objeto SeMiLLa guardado previamente.
}
