% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosechar.R
\name{guardar}
\alias{guardar}
\title{Guardar Objeto SeMiLLa}
\usage{
guardar(semilla, archivo = "semilla", verbose = TRUE)
}
\arguments{
\item{semilla}{Objeto semilla}

\item{archivo}{Nombre del archivo (sin extension .rds)}

\item{verbose}{Mostrar progreso}
}
\description{
Guarda el objeto completo para uso posterior.
}
