% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases_v2.R
\name{homogeneidad_semantica}
\alias{homogeneidad_semantica}
\title{Homogeneidad semantica (alias de fiabilidad_semantica)}
\usage{
homogeneidad_semantica(x, metodo = "spearman_brown", verbose = TRUE)
}
\arguments{
\item{x}{Objeto semilla con embeddings, o matriz de similitud}

\item{metodo}{Metodo de calculo: "spearman_brown"}

\item{verbose}{Mostrar progreso}
}
\value{
Igual que \code{\link{fiabilidad_semantica}}.
}
\description{
Alias descriptivo de \code{\link{fiabilidad_semantica}}. Resume la
homogeneidad tipo alfa de cada dimension a partir de la similitud media. Es
un proxy pre-empirico de consistencia, no fiabilidad poblacional.
}
\seealso{
\code{\link{fiabilidad_semantica}}, \code{\link{auditar_redundancia}}
}
