% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germinar.R
\name{items_similares}
\alias{items_similares}
\title{Buscar Items Similares}
\usage{
items_similares(embeddings, item, top = 5)
}
\arguments{
\item{embeddings}{Objeto semilla_embeddings o semilla}

\item{item}{Numero del item o texto}

\item{top}{Numero de items similares (default: 5)}
}
\value{
Dataframe con items similares
}
\description{
Encuentra items semanticamente similares a un item dado.
}
\examples{
\dontrun{
# Items similares al item 1
items_similares(emb, item = 1, top = 5)
}

}
