% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{limpiar_cache}
\alias{limpiar_cache}
\title{Limpiar Cache}
\usage{
limpiar_cache(confirmar = TRUE)
}
\arguments{
\item{confirmar}{Si TRUE (default), pide confirmacion}
}
\description{
Elimina todas las entradas del cache actual.
}
