% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_embeddings}
\alias{plot_embeddings}
\title{Visualizar Embeddings en 2D}
\usage{
plot_embeddings(
  x,
  metodo = "tsne",
  colorear_por = "dimension",
  mostrar_etiquetas = TRUE,
  perplexity = NULL,
  semilla = 42
)
}
\arguments{
\item{x}{Objeto semilla o semilla_embeddings}

\item{metodo}{Metodo de reduccion: "tsne" (default) o "umap"}

\item{colorear_por}{Variable para colorear: "dimension" o "factor_efa"}

\item{mostrar_etiquetas}{Mostrar numero de item}

\item{perplexity}{Perplexity para t-SNE (default: 5 o n/4)}

\item{semilla}{Semilla para reproducibilidad}
}
\value{
Objeto ggplot2
}
\description{
Proyecta los embeddings de alta dimension a 2D usando t-SNE o UMAP,
coloreando los items por dimension o factor.
}
\examples{
\dontrun{
# Proyeccion t-SNE
plot_embeddings(mi_escala)

# Colorear por factor EFA
plot_embeddings(mi_escala, colorear_por = "factor_efa")
}

}
