% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_evolucion_precision}
\alias{plot_evolucion_precision}
\title{Grafico de Evolucion de Precision (Refinamiento)}
\usage{
plot_evolucion_precision(datos_iteraciones, objetivo = 90, titulo = NULL)
}
\arguments{
\item{datos_iteraciones}{Dataframe con columnas: Iteracion, Precision}

\item{objetivo}{Precision objetivo (linea de referencia)}

\item{titulo}{Titulo del grafico (default: auto)}
}
\value{
Objeto ggplot2
}
\description{
Muestra la evolucion de la precision a traves de iteraciones de refinamiento.
Util para documentar el proceso de mejora de la escala.
}
\examples{
\dontrun{
# Despues de refinamiento iterativo
datos <- data.frame(
  Iteracion = 1:3,
  Precision = c(80, 85, 95)
)
plot_evolucion_precision(datos, objetivo = 90)
}

}
