% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fiabilidad}
\alias{plot_fiabilidad}
\title{Visualizar Fiabilidad por Dimension}
\usage{
plot_fiabilidad(fiab)
}
\arguments{
\item{fiab}{Objeto semilla_fiabilidad de fiabilidad_semantica()}
}
\value{
Objeto ggplot2
}
\description{
Grafico de barras mostrando alpha semantico por dimension con
interpretacion de niveles.
}
