% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_forma_corta}
\alias{plot_forma_corta}
\title{Visualizar Forma Corta}
\usage{
plot_forma_corta(fc, escala_original = NULL)
}
\arguments{
\item{fc}{Resultado de forma_corta()}

\item{escala_original}{Objeto semilla original (opcional)}
}
\value{
Objeto ggplot2
}
\description{
Compara la distribucion de items entre escala original y forma corta.
}
