% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_precision}
\alias{plot_precision}
\title{Grafico de Precision por Dimension}
\usage{
plot_precision(x, titulo = NULL)
}
\arguments{
\item{x}{Objeto semilla_precision (resultado de precision_clasificacion)}

\item{titulo}{Titulo del grafico (default: auto)}
}
\value{
Objeto ggplot2
}
\description{
Genera un grafico de barras horizontales mostrando la precision de
clasificacion para cada dimension. Incluye colores segun nivel de precision.
}
\examples{
\dontrun{
prec <- precision_clasificacion(mi_escala)
plot_precision(prec)
}

}
