% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scree}
\alias{plot_scree}
\title{Scree Plot con Parallel Analysis}
\usage{
plot_scree(x)
}
\arguments{
\item{x}{Objeto semilla o semilla_efa}
}
\value{
Objeto ggplot2
}
\description{
Visualiza eigenvalues del EFA junto con la linea de parallel analysis
para determinar el numero optimo de factores.
}
