% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_v_aiken}
\alias{plot_v_aiken}
\title{Visualizar V de Aiken por Item}
\usage{
plot_v_aiken(cv, ordenar = TRUE, corte = 0.7)
}
\arguments{
\item{cv}{Objeto semilla_cv de validez_contenido()}

\item{ordenar}{Si TRUE, ordena por V de Aiken}

\item{corte}{Valor de corte (default: 0.70)}
}
\value{
Objeto ggplot2
}
\description{
Grafico lollipop mostrando V de Aiken para cada item con intervalos
de confianza y linea de corte.
}
