% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semilla.R
\name{print.semilla}
\alias{print.semilla}
\title{Imprimir objeto SeMiLLa}
\usage{
\method{print}{semilla}(x, ...)
}
\arguments{
\item{x}{Objeto de clase semilla}

\item{...}{Argumentos adicionales}
}
\description{
Imprimir objeto SeMiLLa
}
