% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluar.R
\name{print.semilla_cv}
\alias{print.semilla_cv}
\title{Print Content Validity Results (V de Aiken)}
\usage{
\method{print}{semilla_cv}(x, ...)
}
\arguments{
\item{x}{Objeto semilla_cv}

\item{...}{Argumentos adicionales}
}
\description{
Print Content Validity Results (V de Aiken)
}
