% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluar.R
\name{print.semilla_fiabilidad}
\alias{print.semilla_fiabilidad}
\title{Print Semantic Reliability Results}
\usage{
\method{print}{semilla_fiabilidad}(x, ...)
}
\arguments{
\item{x}{Objeto semilla_fiabilidad}

\item{...}{Argumentos adicionales}
}
\description{
Print Semantic Reliability Results
}
