% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sembrar.R
\name{ver_items}
\alias{ver_items}
\title{Ver Items Generados}
\usage{
ver_items(x, dimension = NULL)
}
\arguments{
\item{x}{Objeto semilla, semilla_items, o dataframe}

\item{dimension}{Filtrar por dimension (NULL = todas)}
}
\value{
Dataframe con columnas: factor, item
}
\description{
Devuelve un dataframe con los items organizados por factor.
}
\examples{
\dontrun{
# Ver todos los items
ver_items(mi_escala)

# Ver items de una dimension
ver_items(mi_escala, dimension = "autoeficacia")
}

}
